+++ runtime/service/src/test/java/org/apache/polaris/service/catalog/iceberg/LocalIcebergCatalogTest.java
@@ -109,7 +109,8 @@ class LocalIcebergCatalogTest {
             null,
             null,
             null,
-            null);
+            null,
+            new TableMetadataCache(TestTableMetadataCacheConfiguration.disabled()));
   }
 
   @Test