+++ runtime/service/src/testFixtures/java/org/apache/polaris/service/TestServices.java
@@ -99,6 +99,8 @@ import org.apache.polaris.service.catalog.iceberg.IcebergRestConfigEndpoints;
 import org.apache.polaris.service.catalog.iceberg.IcebergRestConfigurationEventServiceDelegator;
 import org.apache.polaris.service.catalog.iceberg.IcebergViewConfigEndpoints;
 import org.apache.polaris.service.catalog.iceberg.ImmutableIcebergCatalogHandler;
+import org.apache.polaris.service.catalog.iceberg.TableMetadataCache;
+import org.apache.polaris.service.catalog.iceberg.TestTableMetadataCacheConfiguration;
 import org.apache.polaris.service.catalog.io.FileIOFactory;
 import org.apache.polaris.service.catalog.io.MeasuredFileIOFactory;
 import org.apache.polaris.service.catalog.io.StorageAccessConfigProvider;
@@ -403,6 +405,7 @@ public record TestServices(
               metaStoreManager,
               callContext,
               principal,
+              new TableMetadataCache(TestTableMetadataCacheConfiguration.disabled()),
               idempotencyRequestContext);
 
       ReservedProperties reservedProperties = ReservedProperties.NONE;